Export texture on VRM materials' demand
UniGLTF exports only main textures on Unity Player, but VRM materials can
refer to more textures. Exporting texture on VRM materials' demand is so
naive but less intrusive for UniGLTF and thus reasonable in this case.

Now, CreateFromMaterial static method of glTF_VRM_Material is replaced with
Export static method which takes glTF and adds glTF_VRM_Material to the
glTf with a side effect to add gltfTexture.